+++ repository/session_repository.go
@@ -2,7 +2,6 @@ package repository
 
 import (
 	"context"
-	"database/sql"
 	"github.com/google/uuid"
 	sqlc "readly/db/sqlc"
 	"time"
@@ -25,34 +24,8 @@ func NewSessionRepository(q sqlc.Querier) SessionRepository {
 	}
 }
 
-type CreateSessionRequest struct {
-	ID           uuid.UUID
-	UserID       int64
-	RefreshToken string
-	ExpiresAt    time.Time
-	IPAddress    string
-	UserAgent    string
-}
-
 func (r *SessionRepositoryImpl) CreateSession(ctx context.Context, req CreateSessionRequest) error {
-	ipAddress := sql.NullString{String: "", Valid: false}
-	if req.IPAddress != "" {
-		ipAddress = sql.NullString{String: req.IPAddress, Valid: true}
-	}
-	userAgent := sql.NullString{String: "", Valid: false}
-	if req.UserAgent != "" {
-		userAgent = sql.NullString{String: req.UserAgent, Valid: true}
-	}
-
-	args := sqlc.CreateSessionParams{
-		ID:           req.ID,
-		UserID:       req.UserID,
-		RefreshToken: req.RefreshToken,
-		ExpiresAt:    req.ExpiresAt,
-		IpAddress:    ipAddress,
-		UserAgent:    userAgent,
-	}
-	_, err := r.querier.CreateSession(ctx, args)
+	_, err := r.querier.CreateSession(ctx, req.toSQLC())
 	if err != nil {
 		return err
 	}
